ADO2 - REFATORACAO

+++ senac-cn-ado02/lib.h
@@ -1,4 +1,4 @@
-int fatorial(int n);
+double fatorial(double n);
 
 double cosseno(float x);
